Added checking for dependencies to the ProductService.DeleteAsync method. If product has any dependencies in PayingItemProduct table then it cannot be deleted. Removed unused directives from Product model. Removed references to Irepository<Product> from ProductController and injected IProductService instead of it.

+++ BussinessLogic/Services/ProductService.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using System.Linq;
 using System.Threading.Tasks;
 using HomeAccountingSystem_DAL.Model;
 using HomeAccountingSystem_DAL.Repositories;
@@ -38,6 +39,11 @@ namespace BussinessLogic.Services
 
         public async Task DeleteAsync(int id)
         {
+            var dependencies = (await _productRepository.GetItemAsync(id)).PaiyngItemProduct.Any();
+            if (dependencies)
+            {
+                return;
+            }
             await _productRepository.DeleteAsync(id);
             await _productRepository.SaveAsync();
         }
